Adding node demand multiplier to Wotan.
Previously node demands were scaled by adjusting opin demands. But now that
paths can be analyzed out of ipins, scaling via opin demand is no longer sufficient.
The demand multiplier is a scaling factor that is applied to the demands of all nodes
with the exception of SOURCE/SINK/IPIN/OPIN.